SMART 2 dynamised: works HTML creation is prepared.
